Code-Samples/Python/requirements.txt
# Azure AI Models Course - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Azure AI Services
azure-ai-textanalytics>=5.3.0
azure-cognitiveservices-vision-computervision>=0.9.0
azure-cognitiveservices-speech>=1.34.0
azure-ai-formrecognizer>=3.3.0

# Azure OpenAI
openai>=1.52.0

# Azure Core Libraries
azure-core>=1.29.5
azure-identity>=1.15.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
requests>=2.31.0
pillow>=10.1.0
numpy>=1.24.0
pandas>=2.1.0

# Audio Processing (for Speech samples)
pydub>=0.25.1
wave>=0.0.2

# Web Framework (for API examples)
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0

# Development Tools
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0

# Jupyter Notebooks (optional)
jupyter>=1.0.0
ipywidgets>=8.1.0

# Documentation
markdown>=3.5.0 